Refactor notification handling and various loops.
This _seems_ to be more reliable, but needs more testing.

+++ lib/furble/Device.h
@@ -26,9 +26,9 @@ class Device {
    * GPS data type.
    */
   typedef struct _gps_t {
-    float latitude;
-    float longitude;
-    float altitude;
+    double latitude;
+    double longitude;
+    double altitude;
   } gps_t;
 
   /**
